# openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "description": "分析请求",
        "examples": [
          {
            "data": {
              "energy_cost": 1.2,
              "financing_cost": 4.5,
              "land_price": 800.0,
              "salary_level": 15000.0
            },
            "industry": "半导体",
            "region": "深圳",
            "year": 2025
          }
        ],
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "industry": {
            "default": "半导体",
            "title": "Industry",
            "type": "string"
          },
          "region": {
            "default": "深圳",
            "title": "Region",
            "type": "string"
          },
          "year": {
            "default": 2025,
            "description": "分析年份",
            "maximum": 2200.0,
            "minimum": 1900.0,
            "title": "Year",
            "type": "integer"
          }
        },
        "title": "AnalysisRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "中国城市经济数据 REST API：35 城、16 年指标、集成预测与企业/政府/产业分析。",
    "license": {
      "name": "GPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "title": "Urban Pulse",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analysis/enterprise": {
      "post": {
        "description": "基于真实城市数据，对企业选址进行成本、供应链、政策环境评分。",
        "operationId": "enterprise_analysis_api_v1_analysis_enterprise_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "企业端综合分析",
        "tags": [
          "企业端"
        ]
      }
    },
...
    "/api/v1/analysis/enterprise/compare": {
      "post": {
        "description": "对比多个城市的企业选址友好度。",
        "operationId": "compare_enterprises_api_v1_analysis_enterprise_compare_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "title": "City Names",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Enterprises Api V1 Analysis Enterprise Compare Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "多城市企业选址对比",
        "tags": [
          "企业端"
        ]
      }
    },
    "/api/v1/analysis/enterprise/location/{city_name}": {
      "get": {
        "description": "分析指定城市的企业选址友好度。",
        "operationId": "analyze_enterprise_location_api_v1_analysis_enterprise_location__city_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "city_name",
            "required": true,
            "schema": {
              "title": "City Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Analyze Enterprise Location Api V1 Analysis Enterprise Location  City Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "企业选址分析",
        "tags": [
          "企业端"
        ]
      }
    },
    "/api/v1/analysis/enterprise/sample": {
      "get": {
        "description": "返回企业端分析所需的指标模板，可在此基础上修改后提交分析。",
        "operationId": "get_enterprise_sample_data_api_v1_analysis_enterprise_sample_get",
        "parameters": [
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "default": "深圳",
              "title": "Region",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "industry",
            "required": false,
            "schema": {
              "default": "半导体",
              "title": "Industry",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "default": 2025,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取企业端示例数据",
        "tags": [
          "企业端"
        ]
      }
    },
    "/api/v1/analysis/government": {
      "post": {
        "description": "基于真实城市数据，评估财政杠杆、产业带动效应与产业链完整性。",
        "operationId": "government_analysis_api_v1_analysis_government_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "政府端产业分析",
        "tags": [
          "企业端",
          "政府端"
        ]
      }
    },
...
  },
...
}